requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
tqdm
python-dotenv
google-genai
transformers
torch
